refactor(beamline-states): BeamlineStateManager widget moved to separate module
